+++ packages/orq-rc/src/orq_ai_sdk/models/bulkcreatedatapointsop.py
@@ -471,5 +471,5 @@ class ResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-28T11:21:06.214Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-29T12:34:18.367Z")
     r"""The date and time the resource was last updated"""