+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-05-28T11:21:08.972Z")
+    created: Optional[datetime] = parse_datetime("2025-05-29T12:34:21.136Z")
     r"""The date and time the resource was created"""