+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JWB9ZZ2X2PJVJERART1X6A77"
+        "01JWE0JQ8E7P71NHV2H1D5899R"
     )
     r"""The id of the resource"""