+++ packages/orq-rc/RELEASES.md
@@ -1579,3 +1579,13 @@ Based on:
 - [python v3.6.0-rc.43] packages/orq-rc
 ### Releases
 - [PyPI v3.6.0-rc.43] https://pypi.org/project/orq-ai-sdk/3.6.0-rc.43 - packages/orq-rc
+
+## 2025-05-29 12:35:44
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.552.0 (2.610.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.7.0-rc.1] packages/orq-rc
+### Releases
+- [PyPI v3.7.0-rc.1] https://pypi.org/project/orq-ai-sdk/3.7.0-rc.1 - packages/orq-rc